Refactor create function
